chore: rename gsheets-helpers to a1-notation-utils
